+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/mapper/AllergyIntoleranceMapper.java
@@ -24,6 +24,7 @@ import org.hl7.fhir.dstu3.model.CodeableConcept;
 import org.hl7.fhir.dstu3.model.DateTimeType;
 import org.hl7.fhir.dstu3.model.Encounter;
 import org.hl7.fhir.dstu3.model.Extension;
+import org.hl7.fhir.dstu3.model.Meta;
 import org.hl7.fhir.dstu3.model.Patient;
 import org.hl7.fhir.dstu3.model.Reference;
 import org.hl7.v3.CD;
@@ -40,6 +41,7 @@ import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
 import uk.nhs.adaptors.pss.translator.util.DegradedCodeableConcepts;
 import uk.nhs.adaptors.pss.translator.util.ResourceFilterUtil;
+import uk.nhs.adaptors.pss.translator.util.builder.MetaBuilder;
 
 @Service
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
@@ -88,6 +90,13 @@ public class AllergyIntoleranceMapper extends AbstractMapper<AllergyIntolerance>
                 .getId()
                 .getRoot();
 
+        final Meta meta = new MetaBuilder()
+            .withInitialMeta(() -> generateMeta(META_PROFILE))
+            .withSecurityIfConfidentialityCodesPresent(
+                ehrComposition.getConfidentialityCode(),
+                observationStatement.getConfidentialityCode()
+            ).build();
+
         allergyIntolerance
             .addCategory(getCategory(compoundStatement))
             .setAssertedDateElement(getAssertedDateElement(compoundStatement.getAvailabilityTime(), ehrComposition))
@@ -95,7 +104,7 @@ public class AllergyIntoleranceMapper extends AbstractMapper<AllergyIntolerance>
             .setClinicalStatus(ACTIVE)
             .setVerificationStatus(UNCONFIRMED)
             .addIdentifier(buildIdentifier(id, practiseCode))
-            .setMeta(generateMeta(META_PROFILE))
+            .setMeta(meta)
             .setId(id);
 
         buildOnset(compoundStatement, allergyIntolerance);