+++ gp2gp-translator/src/test/java/uk/nhs/adaptors/pss/translator/mapper/AllergyIntoleranceMapperTest.java
@@ -15,17 +15,20 @@ import static org.springframework.util.ResourceUtils.getFile;
 
 import static uk.nhs.adaptors.pss.translator.util.XmlUnmarshallUtil.unmarshallFile;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
 
 import org.hl7.fhir.dstu3.model.AllergyIntolerance;
 import org.hl7.fhir.dstu3.model.CodeableConcept;
+import org.hl7.fhir.dstu3.model.Coding;
 import org.hl7.fhir.dstu3.model.Encounter;
 import org.hl7.fhir.dstu3.model.Extension;
 import org.hl7.fhir.dstu3.model.Identifier;
 import org.hl7.fhir.dstu3.model.Patient;
 import org.hl7.fhir.dstu3.model.Reference;
 import org.hl7.v3.CD;
+import org.hl7.v3.CV;
 import org.hl7.v3.RCMRMT030101UKEhrExtract;
 
 import org.junit.jupiter.api.Test;
@@ -45,7 +48,7 @@ import uk.nhs.adaptors.pss.translator.util.DegradedCodeableConcepts;
 import static uk.nhs.adaptors.common.util.CodeableConceptUtils.createCodeableConcept;
 
 @ExtendWith(MockitoExtension.class)
-public class AllergyIntoleranceMapperTest {
+class AllergyIntoleranceMapperTest {
 
     private static final String XML_RESOURCES_BASE = "xml/AllergyIntolerance/";
     private static final String COMPOUND_STATEMENT_ROOT_ID = "394559384658936";
@@ -74,6 +77,10 @@ public class AllergyIntoleranceMapperTest {
     private static final String MULTILEX_COCONUT_OIL = "01142009";
     private static final String SNOMED_CODE_SYSTEM = "2.16.840.1.113883.2.1.3.2.4.15";
     private static final String SNOMED_COCONUT_OIL = "14613911000001107";
+    private static final Coding CONFIDENTIALITY_CODING = new Coding()
+        .setSystem("http://hl7.org/fhir/v3/ActCode")
+        .setCode("NOPAT")
+        .setDisplay("no disclosure to patient, family or caregivers without attending provider's authorization");
 
     @Mock
     private CodeableConceptMapper codeableConceptMapper;
@@ -85,20 +92,22 @@ public class AllergyIntoleranceMapperTest {
     private AllergyIntoleranceMapper allergyIntoleranceMapper;
 
     @Test
-    public void testMapDrugAllergyWithAllData() {
+    void testGivenDrugAllergyWithAllDataThenAllDataPopulated() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("drug-allergy-structure.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
             .thenReturn(defaultCodeableConcept())
             .thenReturn(secondaryCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("drug-allergy-structure.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
             getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertFixedValues(allergyIntolerance);
-
         assertExtension(allergyIntolerance);
+
         assertThat(allergyIntolerance.getCategory().get(0).getValue()).isEqualTo(MEDICATION);
         assertThat(allergyIntolerance.getAssertedDateElement().asStringValue()).isEqualTo("1978-12-31");
         assertThat(allergyIntolerance.getRecorder().getReference()).isEqualTo("Practitioner/2D70F602-6BB1-47E0-B2EC-39912A59787D");
@@ -113,17 +122,18 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void testMapAuthorAndParticipantToRecorderAndAsserterAllergyIntolerance() {
+    void testGivenAuthorAndParticipantThenMapsToRecorderAndAsserterAllergyIntolerance() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-valid-author-and-participant2.xml");
 
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
             .thenReturn(defaultCodeableConcept())
             .thenReturn(secondaryCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("allergy-structure-with-valid-author-and-participant2.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
-                                                                                             getEncounterList(), PRACTISE_CODE);
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+            getEncounterList(), PRACTISE_CODE);
 
         assertEquals(1, allergyIntolerances.size());
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertAll(
             () -> assertExtension(allergyIntolerance),
@@ -132,19 +142,23 @@ public class AllergyIntoleranceMapperTest {
         );
     }
 
+    /**
+     * At this moment it is not very clear if this is the correct behavior.
+     * We haven't seen a supplier send over a HL7 in this form, but we want to specify some behaviour.
+     */
     @Test
-    public void testGivenAuthorAndAutParticipant2AuthorAndRecorderPopulatedWithParticipant2() {
-        // At this moment it is not very clear if this is the correct behavior.
-        // We haven't seen a supplier send over a HL7 in this form, but we want to specify some behaviour.
+    void testGivenAuthorAndAutParticipant2AuthorAndRecorderThenPopulatedWithParticipant2() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-participant-of-aut-typecode.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
             .thenReturn(defaultCodeableConcept())
             .thenReturn(secondaryCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("allergy-structure-with-participant-of-aut-typecode.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
-                                                                                             getEncounterList(), PRACTISE_CODE);
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+            getEncounterList(), PRACTISE_CODE);
 
         assertEquals(1, allergyIntolerances.size());
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertAll(
             () -> assertExtension(allergyIntolerance),
@@ -154,18 +168,107 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void testGivenAuthorAndMultipleParticipant2sAndOneAutParticipant2AuthorAndRecorderPopulatedWithAuthorAndParticipant2() {
-        // At this moment it is not very clear if this is the correct behavior with such number of participants.
-        // We haven't seen a supplier send over a HL7 in this form, but we want to specify some behaviour.
+    void testGivenConfidentialityCodeWithNopatWithinObservationStatementThenMetaSecurityPopulated() {
+        final RCMRMT030101UKEhrExtract ehrExtract =
+            unmarshallEhrExtract("allergy-structure-with-observation-statement-confidentiality-code.xml");
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper
+            .mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
+
+        assertEquals(1, allergyIntolerances.size());
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
+        assertThat(allergyIntolerance.getMeta().getSecurity())
+            .usingRecursiveComparison()
+            .isEqualTo(Collections.singletonList(CONFIDENTIALITY_CODING));
+    }
+
+    @Test
+    void testGivenConfidentialityCodeWithNopatWithinEhrCompositionAndNotObservationStatementThenMetaSecurityPopulated() {
+        final RCMRMT030101UKEhrExtract ehrExtract =
+            unmarshallEhrExtract("allergy-structure-with-ehr-composition-confidentiality-code.xml");
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper
+            .mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
+
+        assertEquals(1, allergyIntolerances.size());
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
+        assertThat(allergyIntolerance.getMeta().getSecurity())
+            .usingRecursiveComparison()
+            .isEqualTo(Collections.singletonList(CONFIDENTIALITY_CODING));
+    }
+
+    @Test
+    void testGivenConfidentialityCodeWithCodeOtherThanNopatWithinObservationStatementThenMetaSecurityNotPopulated() {
+        final RCMRMT030101UKEhrExtract ehrExtract =
+            unmarshallEhrExtract("allergy-structure-with-observation-statement-confidentiality-code.xml");
+        final CV invalidCoding = new CV();
+
+        invalidCoding.setCode("NOSCRUB");
+
+        ehrExtract
+            .getComponent().get(0).getEhrFolder()
+            .getComponent().get(0).getEhrComposition()
+            .getComponent().get(0).getCompoundStatement()
+            .getComponent().get(0).getObservationStatement()
+            .setConfidentialityCode(invalidCoding);
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper
+            .mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
+
+        assertEquals(1, allergyIntolerances.size());
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
+        assertThat(allergyIntolerance.getMeta().getSecurity()).isEmpty();
+    }
+
+    @Test
+    void testGivenConfidentialityCodeWithCodeOtherThanNopatWithinEhrCompositionThenMetaSecurityNotPopulated() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-ehr-composition-confidentiality-code.xml");
+        final CV invalidCoding = new CV();
+
+        invalidCoding.setCode("NOSCRUB");
+
+        ehrExtract
+            .getComponent().get(0).getEhrFolder()
+            .getComponent().get(0).getEhrComposition()
+            .setConfidentialityCode(invalidCoding);
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper
+            .mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
+
+        assertEquals(1, allergyIntolerances.size());
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
+        assertThat(allergyIntolerance.getMeta().getSecurity()).isEmpty();
+    }
+
+    @Test
+    void testGivenNoConfidentialityCodeThenMetaSecurityNotPopulated() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-participant-of-aut-typecode.xml");
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper
+            .mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
+
+        assertEquals(1, allergyIntolerances.size());
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
+        assertThat(allergyIntolerance.getMeta().getSecurity()).isEmpty();
+    }
+
+    /**
+     * At this moment it is not very clear if this is the correct behavior with such number of participants.
+     * We haven't seen a supplier send over a HL7 in this form, but we want to specify some behaviour.
+     */
+    @Test
+    void testGivenAuthorAndMultipleParticipant2sAndOneAutParticipant2AuthorAndRecorderThenPopulatedWithAuthorAndParticipant2() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-author-and-multiple-participants.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
             .thenReturn(defaultCodeableConcept())
             .thenReturn(secondaryCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("allergy-structure-with-author-and-multiple-participants.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
-                                                                                             getEncounterList(), PRACTISE_CODE);
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+            getEncounterList(), PRACTISE_CODE);
 
         assertEquals(1, allergyIntolerances.size());
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertAll(
             () -> assertExtension(allergyIntolerance),
@@ -175,16 +278,18 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void testMapNonDrugAllergyWithAllData() {
+    void testGivenNonDrugAllergyWithAllDataThenMapsSuccessfully() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("non-drug-allergy-structure.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
             .thenReturn(defaultCodeableConcept())
             .thenReturn(secondaryCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("non-drug-allergy-structure.xml");
-        List<AllergyIntolerance> allergyIntolerances
-                                    = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
+
+        final List<AllergyIntolerance> allergyIntolerances
+            = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
 
         assertEquals(1, allergyIntolerances.size());
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertFixedValues(allergyIntolerance);
 
@@ -201,49 +306,51 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void testMapDegradedNonDrugAllergy() {
-        when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
-                .thenReturn(new CodeableConcept());
+    void testGivenCompoundStatementCodeOfNonDrugAllergyCodeThenSetsCodeToTransferDegradedNonDrugAllergy() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("degraded-non-drug-allergy-structure.xml");
 
-        var ehrExtract = unmarshallEhrExtract("degraded-non-drug-allergy-structure.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+        when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
+            .thenReturn(new CodeableConcept());
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
             getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertThat(allergyIntolerance.getCode().getCodingFirstRep())
             .isEqualTo(DegradedCodeableConcepts.DEGRADED_NON_DRUG_ALLERGY);
     }
 
     @Test
-    public void testMapDegradedDrugAllergy() {
+    void testGivenCompoundStatementCodeOfDrugAllergyCodeThenSetsCodeToTransferDegradedDrugAllergy() {
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
                 .thenReturn(nonSnomedCodeableConcept());
 
-        var ehrExtract = unmarshallEhrExtract("degraded-drug-allergy-structure.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("degraded-drug-allergy-structure.xml");
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
                 getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertThat(allergyIntolerance.getCode().getCodingFirstRep())
                 .isEqualTo(DegradedCodeableConcepts.DEGRADED_DRUG_ALLERGY);
     }
 
     @Test
-    public void When_AllergyWithOriginalTextAndNoValue_Expect_MapsCodingTextFromCodeOriginalText() {
+    void testGivenAllergyWithOriginalTextAndNoValueThenMapsCodingTextFromCodeOriginalText() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-original-text-in-code.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
             .thenReturn(defaultCodeableConcept())
             .thenReturn(secondaryCodeableConcept());
 
-        var ehrExtract = unmarshallEhrExtract("allergy-structure-with-original-text-in-code.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
             getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertFixedValues(allergyIntolerance);
 
@@ -251,17 +358,17 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void testMapAllergyWithNoOptionalData() {
+    void testGivenAllergyWithNoOptionalData() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-optional-data.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class))).thenReturn(defaultCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("allergy-structure-with-optional-data.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
             getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
-
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
         assertFixedValues(allergyIntolerance);
-
         assertThat(allergyIntolerance.getCode().getCodingFirstRep()).isEqualTo(DegradedCodeableConcepts.DEGRADED_DRUG_ALLERGY);
         assertThat(allergyIntolerance.getCode().getCoding().get(1).getDisplay()).isEqualTo(CODING_DISPLAY_1);
         assertThat(allergyIntolerance.getAssertedDateElement().asStringValue()).isEqualTo("2019-07-08T13:35:00+00:00");
@@ -272,14 +379,16 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void testMapAllergyAssertedDateFallbackData() {
+    void testGivenUnkAvailabilityTimeThenAssertedDateIsAuthorTime() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-asserted-date-fallback.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class))).thenReturn(defaultCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("allergy-structure-with-asserted-date-fallback.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
-                getEncounterList(), PRACTISE_CODE);
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+            getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertFixedValues(allergyIntolerance);
 
@@ -287,35 +396,41 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void testMapMultipleAllergies() {
+    void testGivenMultipleAllergiesThenExpectAllToBePresent() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-with-multiple-allergy.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class))).thenReturn(defaultCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("allergy-structure-with-multiple-allergy.xml");
-        List<AllergyIntolerance> allergyIntolerances
-                                = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
+
+        final List<AllergyIntolerance> allergyIntolerances
+            = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
 
         assertEquals(THREE, allergyIntolerances.size());
     }
 
     @Test
-    public void testMapStandaloneAllergy() {
+    void testGivenStandaloneAllergyThenNoExtensionPresent() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("allergy-structure-invalid-encounter-reference.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class))).thenReturn(defaultCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("allergy-structure-invalid-encounter-reference.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
             getEncounterList(), PRACTISE_CODE);
 
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
         assertThat(allergyIntolerance.getExtension()).isEmpty();
     }
 
     @Test
-    public void testMapAllergyWithSameTermTexts() {
+    void testGivenAllergyWithSameTermTexts() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("drug-allergy-structure-with-term-text.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class))).thenReturn(defaultCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("drug-allergy-structure-with-term-text.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
             getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertFixedValues(allergyIntolerance);
 
@@ -333,16 +448,18 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void testMapAllergyWithDrugTermText() {
+    void testMapAllergyWithDrugTermText() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("drug-allergy-structure-with-term-text.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
             .thenReturn(tertiaryCodeableConcept())
             .thenReturn(tertiaryCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("drug-allergy-structure-with-term-text.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
             getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertFixedValues(allergyIntolerance);
 
@@ -360,16 +477,18 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void When_DrugAllergyWithValueElement_Expect_MapsCodingTextFromValueDescription() {
+    void testGivenDrugAllergyWithValueElementThenMapsCodingTextFromValueDescription() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("drug-allergy-with-value.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
             .thenReturn(defaultCodeableConcept())
             .thenReturn(secondaryCodeableConcept());
-        var ehrExtract = unmarshallEhrExtract("drug-allergy-with-value.xml");
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
             getEncounterList(), PRACTISE_CODE);
 
         assertThat(allergyIntolerances).hasSize(1);
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertFixedValues(allergyIntolerance);
 
@@ -379,59 +498,57 @@ public class AllergyIntoleranceMapperTest {
     }
 
     @Test
-    public void When_AllergyIntoleranceWithQualifierAndOriginalText_Expect_NotesContainsEpisodicity() {
+    void testGivenAllergyIntoleranceWithQualifierAndOriginalTextThenNotesContainsEpisodicity() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("drug_allergy_with_qualifier_and_original_text.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
-                .thenReturn(defaultCodeableConcept());
+            .thenReturn(defaultCodeableConcept());
 
-        var ehrExtract = unmarshallEhrExtract("drug_allergy_with_qualifier_and_original_text.xml");
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+            getEncounterList(), PRACTISE_CODE);
 
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
-                getEncounterList(), PRACTISE_CODE);
-
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertThat(allergyIntolerance.getNote().get(0).getText())
                 .isEqualTo(EPISODICITY_WITH_ORIGINAL_TEXT_NOTE_TEXT);
     }
 
     @Test
-    public void When_AllergyIntoleranceWithQualifierAndWithoutOriginalText_Expect_NotesContainsEpisodicity() {
+    void testGivenAllergyIntoleranceWithQualifierAndWithoutOriginalTextThenNotesContainsEpisodicity() {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract("drug_allergy_with_qualifier_without_original_text.xml");
+
         when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
-                .thenReturn(defaultCodeableConcept());
+            .thenReturn(defaultCodeableConcept());
 
-        var ehrExtract = unmarshallEhrExtract("drug_allergy_with_qualifier_without_original_text.xml");
+        final List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
+            getEncounterList(), PRACTISE_CODE);
 
-        List<AllergyIntolerance> allergyIntolerances = allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(),
-                getEncounterList(), PRACTISE_CODE);
-
-        var allergyIntolerance = allergyIntolerances.get(0);
+        final AllergyIntolerance allergyIntolerance = allergyIntolerances.get(0);
 
         assertThat(allergyIntolerance.getNote().get(0).getText())
                 .isEqualTo(EPISODICITY_WITHOUT_ORIGINAL_TEXT_NOTE_TEXT);
     }
 
-
-
     @ParameterizedTest
     @MethodSource("allergyStructuresWithTranslations")
-    public void testTppNamedSchemaInValue(String filename) {
-        when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
-                .thenReturn(tertiaryCodeableConcept());
+    void testTppNamedSchemaInValue(String filename) {
+        final RCMRMT030101UKEhrExtract ehrExtract = unmarshallEhrExtract(filename);
 
-        var ehrExtract = unmarshallEhrExtract(filename);
+        when(codeableConceptMapper.mapToCodeableConcept(any(CD.class)))
+            .thenReturn(tertiaryCodeableConcept());
 
         allergyIntoleranceMapper.mapResources(ehrExtract, getPatient(), getEncounterList(), PRACTISE_CODE);
 
         verify(codeableConceptMapper, times(2)).mapToCodeableConcept(cdCaptor.capture());
 
-        CD cd = cdCaptor.getAllValues().get(1);
+        final CD cd = cdCaptor.getAllValues().get(1);
 
         assertThat(cd.getCode()).isEqualTo(MULTILEX_COCONUT_OIL);
         assertThat(cd.getCodeSystem()).isEqualTo(MULTILEX_CODE_SYSTEM);
         assertThat(cd.getDisplayName()).isEqualTo(CODING_DISPLAY_4);
         assertThat(cd.getTranslation().size()).isOne();
 
-        var translation = cd.getTranslation().get(0);
+        final CD translation = cd.getTranslation().get(0);
 
         assertThat(translation.getCodeSystem()).isEqualTo(SNOMED_CODE_SYSTEM);
         assertThat(translation.getCode()).isEqualTo(SNOMED_COCONUT_OIL);